% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisLabel.R
\name{axisLabel}
\alias{axisLabel}
\title{Make an axis label out of a character string}
\usage{
axisLabel(string)
}
\arguments{
\item{string}{\code{character(1)}.}
}
\value{
\code{character(1)}.
Lowercase.
}
\description{
Make an axis label out of a character string
}
\note{
Updated 2019-07-27.
}
\examples{
axisLabel("mitoVsCoding")
axisLabel("log10GenesPerUMI")
axisLabel("Already A Label")
axisLabel("NASA")
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[stringr:str_to_lower]{stringr::str_to_lower()}}
}
}
