% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel-methods.R
\docType{methods}
\name{camel}
\alias{camel}
\alias{camel,ANY-method}
\alias{camel,character-method}
\alias{camel,factor-method}
\alias{camel,matrix-method}
\alias{camel,data.frame-method}
\alias{camel,DataFrame-method}
\alias{camel,GRanges-method}
\alias{camel,GRangesList-method}
\title{Camel case}
\usage{
\S4method{camel}{ANY}(object, strict = FALSE)

\S4method{camel}{character}(object, strict = FALSE)

\S4method{camel}{factor}(object, strict = FALSE)

\S4method{camel}{matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{data.frame}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{DataFrame}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{GRanges}(object, strict = FALSE)

\S4method{camel}{GRangesList}(object, strict = FALSE)
}
\arguments{
\item{object}{Object.}
}
\description{
Camel case
}
