% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/camelCase-methods.R
\name{camelCase}
\alias{camelCase}
\alias{camelCase,character-method}
\title{Camel case}
\usage{
camelCase(object, ...)

\S4method{camelCase}{character}(object, strict = TRUE, smart = TRUE, names = TRUE, prefix = TRUE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\note{
Updated 2021-08-24.
}
\examples{
data(syntactic, package = "AcidTest")
object <- syntactic$character
camelCase(object)
}
