% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotted-methods.R
\docType{methods}
\name{dotted}
\alias{dotted}
\alias{dotted,ANY-method}
\alias{dotted,character-method}
\alias{dotted,factor-method}
\alias{dotted,matrix-method}
\alias{dotted,Matrix-method}
\alias{dotted,data.frame-method}
\alias{dotted,DataFrame-method}
\alias{dotted,GRanges-method}
\alias{dotted,GRangesList-method}
\title{Dotted case}
\usage{
\S4method{dotted}{ANY}(object)

\S4method{dotted}{character}(object)

\S4method{dotted}{factor}(object)

\S4method{dotted}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{data.frame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dotted}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{dotted}{GRanges}(object)

\S4method{dotted}{GRangesList}(object)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Dotted case
}
\note{
\code{\link[=dotted]{dotted()}} support is provided for matching against base R parameters.
However, it is recommended to avoid using it for variable assignments into
an \code{environment}, as that can introduce conflicts with base functions.
}
\examples{
load(system.file("extdata", "mn.rda", package = "syntactic"))
lapply(mn, dotted)
}
