% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dottedCase-methods.R
\docType{methods}
\name{dottedCase}
\alias{dottedCase}
\alias{dottedCase,ANY-method}
\alias{dottedCase,character-method}
\alias{dottedCase,factor-method}
\alias{dottedCase,matrix-method}
\alias{dottedCase,Matrix-method}
\alias{dottedCase,data.frame-method}
\alias{dottedCase,DataFrame-method}
\alias{dottedCase,GRanges-method}
\alias{dottedCase,GRangesList-method}
\alias{dottedCase,SummarizedExperiment-method}
\alias{dotted}
\title{Dotted case}
\usage{
dottedCase(object, ...)

\S4method{dottedCase}{ANY}(object)

\S4method{dottedCase}{character}(object)

\S4method{dottedCase}{factor}(object)

\S4method{dottedCase}{matrix}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{Matrix}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{DataFrame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{dottedCase}{GRanges}(object)

\S4method{dottedCase}{GRangesList}(object)

\S4method{dottedCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE)

dotted(...)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\note{
\code{\link[=dottedCase]{dottedCase()}} support is provided for matching against base R parameters.
However, it is recommended to avoid using it for variable assignments into
an \code{environment}, as that can introduce conflicts with base functions.
}
\examples{
data(syntactic, package = "acidtest")
lapply(syntactic, dottedCase)
}
