% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kebabCase-methods.R
\name{kebabCase}
\alias{kebabCase}
\alias{kebabCase,character-method}
\title{Kebab case}
\usage{
kebabCase(object, ...)

\S4method{kebabCase}{character}(
  object,
  rename = FALSE,
  recursive = FALSE,
  smart = TRUE,
  prefix = !rename
)
}
\arguments{
\item{object}{Object.}

\item{rename}{\code{logical(1)}.
File rename mode. Requires that input contains files and/or directories
that exist on disk. Both case-sensitive (e.g. Linux) and -insensitive (e.g.
macOS, Winddows) file systems are supported.

The \code{prefix} and \code{smart} arguments are supported in this mode.}

\item{recursive}{\code{logical(1)}.
Should the function recurse into directories?
Only applicable when \code{rename = TRUE}.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Format character strings to use kebab-style formatting, where word boundaries
are defined by dashes/hyphens (e.g. \code{this-is-kebab-case}).
}
\details{
Kebab case is recommended for URL slugs and file names. However, they should
not be used for names in R, since dashes are not valid, and should be
substituted with underscores or periods instead.
}
\note{
Updated 2020-01-27.
}
\examples{
data(syntactic, package = "AcidTest")
object <- syntactic$character
kebabCase(object)
}
