% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Make a plot axis or legend label out of a character string}
\usage{
label(string)
}
\arguments{
\item{string}{\code{character(1)}.}
}
\value{
\code{character(1)}.
Lowercase.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\note{
Updated 2019-07-27.
}
\examples{
label("mitoVsCoding")
label("log10GenesPerUMI")
label("Already A Label")
label("NASA")
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[stringr:str_to_lower]{stringr::str_to_lower()}}
}
}
