% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDimnames.R
\name{makeDimnames}
\alias{makeDimnames}
\title{Make syntactically valid dimnames}
\usage{
makeDimnames(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Both row and columns names will be made syntactically valid.
}
\description{
Make syntactically valid dimnames
}
\note{
Updated 2019-07-19.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$matrix
makeDimnames(object)
}
