% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel-methods.R
\name{makeLabel}
\alias{makeLabel}
\alias{makeLabel,character-method}
\title{Make a plot axis or legend label out of a character string}
\usage{
makeLabel(object, ...)

\S4method{makeLabel}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
Lowercase string.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\note{
Updated 2020-01-27.
}
\examples{
makeLabel("mitoVsCoding")
makeLabel("log10GenesPerUMI")
makeLabel("Already A Label")
makeLabel("NASA")
makeLabel("nGene")
}
\seealso{
\itemize{
\item \code{\link[acidgenerics:makeWords]{makeWords()}}.
\item \code{\link[stringr:case]{stringr::str_to_lower()}}
}
}
