% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWords.R
\name{makeWords}
\alias{makeWords}
\title{Convert syntactic names to words separated by spaces}
\usage{
makeWords(names)
}
\arguments{
\item{names}{\code{character}.}
}
\value{
\code{character}.
Words separated by spaces.
}
\description{
Convert syntactic names to words separated by spaces
}
\note{
Updated 2019-07-27.
}
\examples{
makeWords(names = c(
    "killVMaim",
    "log10GenesPerUMI",
    "mitoVsCoding",
    "words already",
    "NASA"
))
}
