% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{assayNames}{\code{logical(1)}.
FIXME}

\item{colData}{\code{logical(1)}.
FIXME}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}

\item{mcols}{\code{logical(1)}.
FIXME}

\item{metadata}{\code{logical(1)}.
FIXME}

\item{names}{\code{logical(1)}.
Sanitize names.}

\item{object}{Object.}

\item{rowData}{\code{logical(1)}.
FIXME}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{x}{Object.}
}
\description{
Parameters
}
\keyword{internal}
