% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel-methods.R, R/dotted-methods.R,
%   R/snake-methods.R, R/upperCamel-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{camel}}, \code{\link[bioverbs]{dotted}}, \code{\link[bioverbs]{snake}}, \code{\link[bioverbs]{upperCamel}}}
}}

