% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFile.R
\name{renameFile}
\alias{renameFile}
\title{Rename files using syntactically valid conventions}
\usage{
renameFile(x, fun = c("kebabCase", "snakeCase", "camelCase", "upperCamelCase"))
}
\description{
Rename files using syntactically valid conventions
}
\note{
Updated 2021-08-24.
}
