% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/snakeCase-methods.R
\docType{methods}
\name{snakeCase}
\alias{snakeCase}
\alias{snakeCase,character-method}
\alias{snake}
\title{Snake case}
\usage{
snakeCase(object, ...)

\S4method{snakeCase}{character}(object, names = TRUE)

snake(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{names}{\code{logical(1)}.
Sanitize names.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\note{
Updated 2019-09-09.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
snakeCase(object)
}
