% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{syntactic-package}
\alias{syntactic}
\alias{syntactic-package}
\title{syntactic}
\description{
Make syntactically valid names out of character vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/syntactic/}
  \item \url{https://github.com/acidgenomics/r-syntactic/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-syntactic/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
