% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{syntacticRename}
\alias{syntacticRename}
\title{Rename files and/or directories using a syntactic naming function}
\usage{
syntacticRename(
  path,
  recursive = FALSE,
  fun = c("kebabCase", "snakeCase", "camelCase", "upperCamelCase"),
  quiet = FALSE
)
}
\arguments{
\item{path}{\code{character}.
File and/or directory paths.}

\item{fun}{\code{character(1)}.
Function name.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}
}
\value{
\code{character}.
Renamed file paths.
}
\description{
Rename files and/or directories using a syntactic naming function
}
\details{
Intelligently deals with a case-insensitive file system, if necessary.
This is very useful for macOS and Windows.

Our syntactic naming functions can result in changes that only differ in
case, which are problematic on case-insensitive mounts, and require movement
of the files into a temporary file name before the final rename.
}
\note{
Updated 2021-08-24.
}
\examples{
testdir <- file.path(tempdir(), "testdata")
unlink(testdir, recursive = TRUE)
dir.create(testdir)
from <- file.path(testdir, c("helloWorld.txt", "fooBar.R"))
file.create(from)
print(basename(from))
to <- syntacticRename(from)
print(basename(to))
unlink(testdir, recursive = TRUE)
}
