% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperCamel-methods.R
\docType{methods}
\name{upperCamel}
\alias{upperCamel}
\alias{upperCamel,ANY-method}
\alias{upperCamel,character-method}
\alias{upperCamel,factor-method}
\alias{upperCamel,matrix-method}
\alias{upperCamel,data.frame-method}
\alias{upperCamel,DataFrame-method}
\alias{upperCamel,GRanges-method}
\alias{upperCamel,GRangesList-method}
\title{Upper camel case}
\usage{
\S4method{upperCamel}{ANY}(object, strict = FALSE)

\S4method{upperCamel}{character}(object, strict = FALSE)

\S4method{upperCamel}{factor}(object, strict = FALSE)

\S4method{upperCamel}{matrix}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamel}{data.frame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamel}{DataFrame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamel}{GRanges}(object, strict = FALSE)

\S4method{upperCamel}{GRangesList}(object, strict = FALSE)
}
\arguments{
\item{object}{Object.}
}
\description{
Upper camel case
}
