% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperCamelCase-methods.R
\docType{methods}
\name{upperCamelCase}
\alias{upperCamelCase}
\alias{upperCamelCase,ANY-method}
\alias{upperCamelCase,character-method}
\alias{upperCamelCase,factor-method}
\alias{upperCamelCase,matrix-method}
\alias{upperCamelCase,Matrix-method}
\alias{upperCamelCase,data.frame-method}
\alias{upperCamelCase,DataFrame-method}
\alias{upperCamelCase,GRanges-method}
\alias{upperCamelCase,GRangesList-method}
\alias{upperCamelCase,SummarizedExperiment-method}
\alias{upperCamel}
\title{Upper camel case}
\usage{
upperCamelCase(object, ...)

\S4method{upperCamelCase}{ANY}(object, strict = FALSE)

\S4method{upperCamelCase}{character}(object, strict = FALSE)

\S4method{upperCamelCase}{factor}(object, strict = FALSE)

\S4method{upperCamelCase}{matrix}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{Matrix}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{DataFrame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{upperCamelCase}{GRanges}(object, strict = FALSE)

\S4method{upperCamelCase}{GRangesList}(object, strict = FALSE)

\S4method{upperCamelCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names formatted in upper camel case.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\examples{
data(mn, package = "acidtest")
lapply(mn, upperCamelCase)
}
