% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel-methods.R
\docType{methods}
\name{camel}
\alias{camel}
\alias{camel,ANY-method}
\alias{camel,character-method}
\alias{camel,factor-method}
\alias{camel,matrix-method}
\alias{camel,Matrix-method}
\alias{camel,data.frame-method}
\alias{camel,DataFrame-method}
\alias{camel,GRanges-method}
\alias{camel,GRangesList-method}
\alias{camel,SummarizedExperiment-method}
\title{Camel case}
\usage{
\S4method{camel}{ANY}(object, strict = FALSE)

\S4method{camel}{character}(object, strict = FALSE)

\S4method{camel}{factor}(object, strict = FALSE)

\S4method{camel}{matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{Matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{data.frame}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{DataFrame}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camel}{GRanges}(object, strict = FALSE)

\S4method{camel}{GRangesList}(object, strict = FALSE)

\S4method{camel}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\examples{
data(mn, package = "acidtest")
lapply(mn, camel)
}
