% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelCase-methods.R
\docType{methods}
\name{camelCase}
\alias{camelCase}
\alias{camelCase,ANY-method}
\alias{camelCase,character-method}
\alias{camelCase,factor-method}
\alias{camelCase,matrix-method}
\alias{camelCase,Matrix-method}
\alias{camelCase,data.frame-method}
\alias{camelCase,DataFrame-method}
\alias{camelCase,GRanges-method}
\alias{camelCase,GRangesList-method}
\alias{camelCase,SummarizedExperiment-method}
\alias{camel}
\title{Camel case}
\usage{
camelCase(object, ...)

\S4method{camelCase}{ANY}(object, strict = FALSE)

\S4method{camelCase}{character}(object, strict = FALSE)

\S4method{camelCase}{factor}(object, strict = FALSE)

\S4method{camelCase}{matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camelCase}{Matrix}(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

\S4method{camelCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{camelCase}{DataFrame}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

\S4method{camelCase}{GRanges}(object, strict = FALSE)

\S4method{camelCase}{GRangesList}(object, strict = FALSE)

\S4method{camelCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use (lower) camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{thisIsCamelCase}).
}
\details{
Camel case is recommended by Bioconductor for variable and function names.
}
\examples{
data(mn, package = "acidtest")
lapply(mn, camelCase)
}
