% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize-methods.R
\docType{methods}
\name{capitalize}
\alias{capitalize}
\alias{capitalize,character-method}
\title{Capitalize}
\usage{
capitalize(object, ...)

\S4method{capitalize}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with names capitalized.
}
\description{
Capitalize
}
\examples{
capitalize(c("hello", "world"))
}
