% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize-methods.R
\name{capitalize}
\alias{capitalize}
\alias{capitalize,character-method}
\title{Capitalize}
\usage{
capitalize(object, ...)

\S4method{capitalize}{character}(object, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict renaming of words containing multiple capital letters,
including acronyms.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Capitalize
}
\note{
Updated 2020-01-27.
}
\examples{
x <- c("the quick Brown fox", "using AIC for model selection")
capitalize(x, strict = FALSE)
capitalize(x, strict = TRUE)
}
