% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/capitalize-methods.R
\docType{methods}
\name{capitalize}
\alias{capitalize}
\alias{capitalize,character-method}
\title{Capitalize}
\usage{
capitalize(object, ...)

\S4method{capitalize}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object, with names capitalized.
}
\description{
Capitalize
}
\note{
Updated 2019-09-09.
}
\examples{
capitalize(c("hello", "world"))
}
