% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize}
\usage{
capitalize(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{character}.
}
\description{
Capitalize the first letter of each string in a \code{character} vector.
}
\examples{
capitalize(c("hello", "world"))
}
\seealso{
\code{R.utils::capitalize()}.
}
