% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize-methods.R
\docType{methods}
\name{capitalize}
\alias{capitalize}
\alias{capitalize,character-method}
\title{Capitalize}
\usage{
\S4method{capitalize}{character}(object)
}
\arguments{
\item{object}{Object.}
}
\description{
Capitalize
}
\examples{
capitalize(c("hello", "world"))
}
