% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dottedCase-methods.R
\name{dottedCase}
\alias{dottedCase}
\alias{dottedCase,character-method}
\alias{dotted}
\title{Dotted case}
\usage{
dottedCase(object, ...)

\S4method{dottedCase}{character}(object, smart = TRUE, names = TRUE, prefix = TRUE)

dotted(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\note{
\code{\link[=dottedCase]{dottedCase()}} support is provided for matching against base R
parameters. However, it is recommended to avoid using it for variable
assignments into an \code{environment}, as that can introduce conflicts with
base functions.

Updated 2019-12-09.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
dottedCase(object)
}
