% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dottedCase-methods.R
\docType{methods}
\name{dottedCase}
\alias{dottedCase}
\alias{dottedCase,character-method}
\alias{dotted}
\title{Dotted case}
\usage{
dottedCase(object, ...)

\S4method{dottedCase}{character}(object, names = TRUE)

dotted(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{names}{\code{logical(1)}.
Sanitize names.}
}
\value{
Modified object, with names formatted in dotted case.
}
\description{
Dotted case
}
\note{
\code{\link[=dottedCase]{dottedCase()}} support is provided for matching against base R
parameters. However, it is recommended to avoid using it for variable
assignments into an \code{environment}, as that can introduce conflicts with
base functions.

Updated 2019-09-09.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
dottedCase(object)
}
