% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/kebabCase-methods.R
\docType{methods}
\name{kebabCase}
\alias{kebabCase}
\alias{kebabCase,character-method}
\alias{kebab}
\title{Kebab case}
\usage{
kebabCase(object, ...)

\S4method{kebabCase}{character}(object, prefix = TRUE)

kebab(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details. Recommended by default.}
}
\value{
Modified object.
}
\description{
Format character strings to use kebab-style formatting, where word boundaries
are defined by dashes/hyphens (e.g. \code{this-is-kebab-case}).
}
\details{
Kebab case is recommended for URL slugs and file names. However, they should
not be used for names in R, since dashes are not valid, and should be
substituted with underscores or periods instead.
}
\note{
We're not including the additional S4 methods that work on
\code{\link[base:names]{names()}} and/or \code{\link[base:dimnames]{dimnames()}} because dashes
are not syntactically valid for names in R.

Updated 2019-10-07.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
kebabCase(object)
}
