% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kebabCase-methods.R
\docType{methods}
\name{kebabCase}
\alias{kebabCase}
\alias{kebabCase,character-method}
\alias{kebab}
\title{Kebab case}
\usage{
kebabCase(object, ...)

\S4method{kebabCase}{character}(object)

kebab(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Format character strings to use kebab-style formatting, where word boundaries
are defined by dashes/hyphens (e.g. \code{this-is-kebab-case}).
}
\details{
Kebab case is recommended for URL slugs and file names. However, they should
not be used for names in R, since dashes are not valid, and should be
substituted with underscores or periods instead.
}
\note{
We're not including the additional S4 methods that work on
\code{\link[base:names]{names()}} and/or \code{\link[base:dimnames]{dimnames()}} because dashes
are not syntactically valid for names in R.
}
\examples{
kebabCase("this is kebabCase case")
}
\seealso{
\code{\link[base:make.names]{make.names()}}.
}
