% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDimnames-methods.R
\name{makeDimnames}
\alias{makeDimnames}
\alias{makeDimnames,ANY-method}
\title{Make syntactically valid dimnames}
\usage{
makeDimnames(object, ...)

\S4method{makeDimnames}{ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Both row and columns names will be made syntactically valid.
}
\description{
Make syntactically valid dimnames
}
\note{
Updated 2020-01-27.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$matrix
makeDimnames(object)
}
