% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabel.R
\name{makeLabel}
\alias{makeLabel}
\title{Make a plot axis or legend label out of a character string}
\usage{
makeLabel(string)
}
\arguments{
\item{string}{\code{character(1)}.}
}
\value{
\code{character(1)}.
Lowercase.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\note{
Updated 2019-07-31.
}
\examples{
makeLabel("mitoVsCoding")
makeLabel("log10GenesPerUMI")
makeLabel("Already A Label")
makeLabel("NASA")
makeLabel("nGene")
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[stringr:str_to_lower]{stringr::str_to_lower()}}
}
}
