% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make syntactically valid names}
\usage{
makeNames(names, unique = TRUE)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid
   names.  This is coerced to character if necessary.}

\item{unique}{logical; if \code{TRUE}, the resulting elements are
     unique.  This may be desired for, e.g., column names.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting \code{names}, the
underlying data returns unchanged.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{names}, \code{rownames}, and/or \code{colnames} without
modification of the values.
}
\note{
\code{makeNames} sanitizes names using underscores instead of dots, the
convention used by \code{make.names}.
}
\examples{
loadRemoteData(url = file.path(basejumpCacheURL, "mn.rda"))

## character ====
x <- mn$character
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

x <- mn$namedCharacter
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

## factor ====
x <- mn$factor
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

## data.frame ====
x <- datasets::USArrests
dimnames(x)
camel(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
dotted(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
snake(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
upperCamel(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()

## list ====
x <- mn$list
print(x)
camel(x) \%>\% names()
dotted(x) \%>\% names()
snake(x) \%>\% names()
upperCamel(x) \%>\% names()
}
\seealso{
\code{make.names()}.
}
