% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make syntactically valid names}
\usage{
makeNames(names, unique = TRUE)
}
\arguments{
\item{names}{\code{character}.
Character vector to be coerced to syntactically valid names.}

\item{unique}{\code{logical(1)}.
If \code{TRUE}, the resulting elements are unique. Recommended by default, for
syntactically valid names (e.g. column, row names). Note that this is
disabled by default for \code{\link[base:make.names]{make.names()}}.}
}
\value{
\code{character}.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{\link[base:names]{names()}}, \code{\link[base:rownames]{rownames()}},
and/or \code{\link[base:colnames]{colnames()}} without modification of the values.
}
\note{
\code{\link[=makeNames]{makeNames()}} sanitizes names using underscores instead of dots, the
convention used by \code{\link[base:make.names]{make.names()}}.

Updated 2019-09-09.
}
\examples{
data(syntactic, package = "acidtest")
names <- syntactic$character
makeNames(names)
}
\seealso{
\code{\link[base:make.names]{make.names()}}.
}
