% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Make syntactically valid names}
\usage{
makeNames(names, unique = TRUE)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid
   names.  This is coerced to character if necessary.}

\item{unique}{logical; if \code{TRUE}, the resulting elements are
     unique.  This may be desired for, e.g., column names.}

\item{object}{\code{character} or class supporting \code{names}.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting \code{names}, the
underlying data returns unchanged.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{names}, \code{rownames}, and/or \code{colnames} without
modification of the values.
}
\note{
\code{makeNames} sanitizes names using underscores instead of dots, the
convention used by \code{make.names}.
}
\examples{
loadRemoteData(url = file.path(basejumpCacheURL, "mn.rda"))

## character ====
x <- mn$character
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

x <- mn$namedCharacter
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

## factor ====
x <- mn$factor
print(x)
camel(x)
dotted(x)
snake(x)
upperCamel(x)
makeNames(x)

## data.frame ====
x <- datasets::USArrests
dimnames(x)
camel(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
dotted(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
snake(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()
upperCamel(x, rownames = TRUE, colnames = TRUE) \%>\% dimnames()

## list ====
x <- mn$list
print(x)
camel(x) \%>\% names()
dotted(x) \%>\% names()
snake(x) \%>\% names()
upperCamel(x) \%>\% names()
}
\seealso{
\itemize{
\item \code{make.names}.
\item \href{https://cran.r-project.org/package=janitor}{janitor} contains a number of
useful functions that provide similar sanitization support, but isn't
designed to work natively with \href{https://bioconductor.org}{Bioconductor} and
S4 object classes.
\item \href{https://cran.r-project.org/package=lettercase}{lettercase}.
\item \href{https://tazinho.github.io/snakecase}{snakecase}.
}
}
