% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames-methods.R
\name{makeNames}
\alias{makeNames}
\alias{makeNames,character-method}
\alias{makeNames,factor-method}
\title{Make syntactically valid names}
\usage{
makeNames(object, ...)

\S4method{makeNames}{character}(object, smart = TRUE, unique = TRUE)

\S4method{makeNames}{factor}(object, smart = TRUE, unique = TRUE)
}
\arguments{
\item{object}{Object.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{unique}{\code{logical(1)}.
If \code{TRUE}, the resulting elements are unique. Recommended by default, for
syntactically valid names (e.g. column, row names). Note that this is
disabled by default for \code{\link[base:make.names]{make.names()}}.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{\link[base:names]{names()}}, \code{\link[base:colnames]{rownames()}},
and/or \code{\link[base:colnames]{colnames()}} without modification of the values.
}
\details{
Sanitizes names using underscores instead of dots, the convention used by
\code{\link[base:make.names]{make.names()}}.
}
\note{
Updated 2020-07-08.
}
\examples{
data(syntactic, package = "AcidTest")
object <- syntactic$character
makeNames(object)
}
\seealso{
\itemize{
\item \href{https://cs.stanford.edu/people/miles/iso8859.html}{ASCII table}
}
}
