% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames-methods.R
\name{makeNames}
\alias{makeNames}
\alias{makeNames,character-method}
\alias{makeNames,factor-method}
\title{Make syntactically valid names}
\usage{
makeNames(object, ...)

\S4method{makeNames}{character}(object, unique = TRUE)

\S4method{makeNames}{factor}(object, unique = TRUE)
}
\arguments{
\item{object}{Object.}

\item{unique}{\code{logical(1)}.
If \code{TRUE}, the resulting elements are unique. Recommended by default, for
syntactically valid names (e.g. column, row names). Note that this is
disabled by default for \code{\link[base:make.names]{make.names()}}.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
For \code{atomic} vectors, these functions will sanitize the values. Otherwise,
they will set \code{\link[base:names]{names()}}, \code{\link[base:colnames]{rownames()}},
and/or \code{\link[base:colnames]{colnames()}} without modification of the values.
}
\details{
Sanitizes names using underscores instead of dots, the convention used by
\code{\link[base:make.names]{make.names()}}.
}
\note{
Updated 2020-01-27.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
makeNames(object)
}
\seealso{
\code{\link[base:make.names]{make.names()}}.
}
