% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTitle.R
\name{makeTitle}
\alias{makeTitle}
\title{Make a title out of character string}
\usage{
makeTitle(string)
}
\arguments{
\item{string}{\code{character(1)}.}
}
\value{
\code{character(1)}.
British style sentence case.
}
\description{
Doesn't modify strings already containing a space or multi-letter acronym.
}
\note{
Updated 2019-07-27.
}
\examples{
makeTitle("mitoVsCoding")
makeTitle("log10GenesPerUMI")
makeTitle("Already A Title")
makeTitle("NASA")
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[=capitalize]{capitalize()}}.
\item \code{\link[stringr:str_to_sentence]{stringr::str_to_sentence()}}.
}
}
