% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWords-methods.R
\name{makeWords}
\alias{makeWords}
\alias{makeWords,character-method}
\title{Convert syntactic names to words separated by spaces}
\usage{
makeWords(object, ...)

\S4method{makeWords}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Words separated by spaces.
}
\description{
Convert syntactic names to words separated by spaces
}
\note{
Updated 2020-01-27.
}
\examples{
makeWords(c(
    "killVMaim",
    "log10GenesPerUMI",
    "mitoVsCoding",
    "words already",
    "NASA"
))
}
