% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{names}{\code{logical(1)}.
Sanitize names.}

\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{string}{\code{character(1)}.}

\item{x}{Object.}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
