% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{object}{Object.}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}

\item{recursive}{\code{logical(1)}.
Should the function recurse into directories?
Only applicable when \code{rename = TRUE}.}

\item{rename}{\code{logical(1)}.
File rename mode. Requires that input contains files and/or directories
that exist on disk. Both case-sensitive (e.g. Linux) and -insensitive (e.g.
macOS, Winddows) file systems are supported.

The \code{prefix} and \code{smart} arguments are supported in this mode.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{unique}{\code{logical(1)}.
If \code{TRUE}, the resulting elements are unique. Recommended by default, for
syntactically valid names (e.g. column, row names). Note that this is
disabled by default for \code{\link[base:make.names]{make.names()}}.}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
