% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{names}{\code{logical(1)}.
Sanitize names.}

\item{object}{Object.}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details. Recommended by default.}

\item{rename}{\code{logical(1)}.
Enable file rename mode.
Requires that input contains files and/or directories that exist on disk.
Either case-sensitive (e.g. Linux) and -insensitive (e.g. macOS, Winddows)
file systems are supported.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc. Recommended by default.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.
Disabled by default.}

\item{...}{Additional arguments.}
}
\description{
Parameters
}
\keyword{internal}
