% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sentenceCase-methods.R
\name{sentenceCase}
\alias{sentenceCase}
\alias{sentenceCase,character-method}
\title{Sentence case}
\usage{
sentenceCase(object, ...)

\S4method{sentenceCase}{character}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Sentence case
}
\note{
Updated 2019-10-21.
}
\examples{
x <- c("the quick Brown fox", "using AIC for model selection")
sentenceCase(x)
}
