% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake-methods.R
\docType{methods}
\name{snake}
\alias{snake}
\alias{snake,ANY-method}
\alias{snake,character-method}
\alias{snake,factor-method}
\alias{snake,matrix-method}
\alias{snake,Matrix-method}
\alias{snake,data.frame-method}
\alias{snake,DataFrame-method}
\alias{snake,GRanges-method}
\alias{snake,GRangesList-method}
\title{Snake case}
\usage{
\S4method{snake}{ANY}(object)

\S4method{snake}{character}(object)

\S4method{snake}{factor}(object)

\S4method{snake}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{data.frame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{DataFrame}(object, rownames = FALSE, colnames = TRUE)

\S4method{snake}{GRanges}(object)

\S4method{snake}{GRangesList}(object)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}
}
\description{
Snake case
}
