% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/snakeCase-methods.R
\name{snakeCase}
\alias{snakeCase}
\alias{snakeCase,character-method}
\title{Snake case}
\usage{
snakeCase(object, ...)

\S4method{snakeCase}{character}(object, smart = TRUE, names = TRUE, prefix = TRUE)
}
\arguments{
\item{object}{Object.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\note{
Updated 2020-07-08.
}
\examples{
data(syntactic, package = "AcidTest")
object <- syntactic$character
snakeCase(object)
}
