% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/snakeCase-methods.R
\docType{methods}
\name{snake}
\alias{snake}
\alias{snakeCase}
\alias{snakeCase,ANY-method}
\alias{snakeCase,character-method}
\alias{snakeCase,factor-method}
\alias{snakeCase,matrix-method}
\alias{snakeCase,Matrix-method}
\alias{snakeCase,data.frame-method}
\alias{snakeCase,DataFrame-method}
\alias{snakeCase,GRanges-method}
\alias{snakeCase,GRangesList-method}
\alias{snakeCase,SummarizedExperiment-method}
\title{Snake case}
\usage{
snakeCase(object, ...)

\S4method{snakeCase}{ANY}(object)

\S4method{snakeCase}{character}(object)

\S4method{snakeCase}{factor}(object)

\S4method{snakeCase}{matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{Matrix}(object, rownames = FALSE, colnames = TRUE)

\S4method{snakeCase}{data.frame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{snakeCase}{DataFrame}(object, rownames = FALSE,
  colnames = TRUE)

\S4method{snakeCase}{GRanges}(object)

\S4method{snakeCase}{GRangesList}(object)

\S4method{snakeCase}{SummarizedExperiment}(object, rownames = FALSE,
  colnames = TRUE)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply sanitization on row names. This is not generally recommended by
default, since rownames commonly contain gene identifiers that should not
be modified.}

\item{colnames}{\code{logical(1)}.
Apply sanitization on column names. This is generally recommended by
default.}
}
\value{
Modified object, with names formatted in snake case.
}
\description{
Format character strings to use snake-style formatting, where word boundaries
are defined by underscores (e.g. \code{this_is_snake_case}).
}
\details{
Snake case is recommended by the tidyverse for variable and function names.
}
\examples{
data(mn, package = "acidtest")
lapply(mn, snakeCase)
}
