% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{syntactic-package}
\alias{syntactic}
\alias{syntactic-package}
\title{syntactic}
\description{
Make syntactically valid names out of character vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steinbaugh/syntactic/}
  \item Report bugs at \url{https://github.com/steinbaugh/syntactic/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
\keyword{internal}
