% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title}
\alias{title}
\title{Make a title out of character string}
\usage{
title(string)
}
\arguments{
\item{string}{\code{character(1)}.}
}
\value{
\code{character(1)}.
British style sentence case.
}
\description{
Make a title out of character string
}
\note{
Updated 2019-07-27.
}
\examples{
title("mitoVsCoding")
title("log10GenesPerUMI")
title("Already A Title")
title("NASA")
}
\seealso{
\itemize{
\item \code{\link[=makeWords]{makeWords()}}.
\item \code{\link[=capitalize]{capitalize()}}.
\item \code{\link[stringr:str_to_sentence]{stringr::str_to_sentence()}}.
}
}
