% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upperCamelCase-methods.R
\docType{methods}
\name{upperCamelCase}
\alias{upperCamelCase}
\alias{upperCamelCase,character-method}
\alias{upperCamel}
\title{Upper camel case}
\usage{
upperCamelCase(object, ...)

\S4method{upperCamelCase}{character}(object, names = TRUE,
  strict = FALSE)

upperCamel(...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{names}{\code{logical(1)}.
Sanitize names.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}
}
\value{
Modified object.
Contains syntatically valid names. For objects supporting
\code{\link[base:names]{names()}}, the underlying data returns unchanged, except for
\code{character} or \code{vector} class.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\note{
Updated 2019-09-09.
}
\examples{
data(syntactic, package = "acidtest")
object <- syntactic$character
upperCamelCase(object)
}
