% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upperCamelCase-methods.R
\name{upperCamelCase}
\alias{upperCamelCase}
\alias{upperCamelCase,character-method}
\title{Upper camel case}
\usage{
upperCamelCase(object, ...)

\S4method{upperCamelCase}{character}(
  object,
  rename = FALSE,
  recursive = FALSE,
  smart = TRUE,
  strict = TRUE,
  names = !rename,
  prefix = !rename
)
}
\arguments{
\item{object}{Object.}

\item{rename}{\code{logical(1)}.
File rename mode. Requires that input contains files and/or directories
that exist on disk. Both case-sensitive (e.g. Linux) and -insensitive (e.g.
macOS, Winddows) file systems are supported.

The \code{prefix} and \code{smart} arguments are supported in this mode.}

\item{recursive}{\code{logical(1)}.
Should the function recurse into directories?
Only applicable when \code{rename = TRUE}.}

\item{smart}{\code{logical(1)}.
Handle complicated special cases, such as mixed case acronyms, plus/minus,
percentages, etc.}

\item{strict}{\code{logical(1)}.
Enforce strict name sanitization. When \code{TRUE}, this does not allow the
return of any capitalized acronyms. "RNA" will become "Rna", for example.}

\item{names}{\code{logical(1)}.
Sanitize names.

Only applies to string mode (\code{rename = FALSE}).}

\item{prefix}{\code{logical(1)}.
Prepend "X" character if necessary, when string begins with a syntactically
invalid character, such as a number or non-alphanumeric. Note that names
are always made syntactically valid when applicable with "X" prefix.
See \code{\link[base:make.names]{make.names()}} for details.

Enabled by default for string mode, but disabled by default for rename
mode, when applicable.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains syntatically valid names.
For objects with \code{\link[base:names]{names()}} defined, the underlying data
returns unchanged, except for \code{character} or \code{vector} class.
}
\description{
Format character strings to use upper camel-style formatting, where word
boundaries are defined by capitlization only (e.g. \code{ThisIsCamelCase}).
}
\details{
Note that lower camel case is generally recommended in R over the use of
upper camel case. However, upper camel case is recommended by Bioconductor
for S4 class names and corresponding generators, but not variables or
functions.
}
\note{
Updated 2021-01-21.
}
\examples{
data(syntactic, package = "AcidTest")
object <- syntactic$character
upperCamelCase(object)
}
